Move chat_archiver atomic_write() to common for re-use
